+++ cypress/e2e/bridged_utils.cy.js
@@ -15,7 +15,7 @@ function initMain(name, version) {
 }
 
 const SDK_NAME = "javascript_native_web";
-const SDK_VERSION = "24.4.1";
+const SDK_VERSION = "24.11.0";
 
 // tests
 describe("Bridged SDK Utilities Tests", () => {